Mostly working database schema
