+++ app/models/__init__.py
@@ -1,0 +1,8 @@
+"""
+Models package - import individual models as needed.
+"""
+
+def configure_relationships():
+    """Call this after importing models to resolve relationships"""
+    from sqlalchemy.orm import configure_mappers
+    configure_mappers()
